+++ http-server/internal/pkg/service/TestService.thrift
