+++ http-server/internal/pkg/service/service.go
@@ -1,4 +1,4 @@
-package config
+package service
 
 import (
 	"io/fs"
